+++ rmcs_ws/src/rmcs_core/src/controller/chassis/chassis_climber_controller.cpp
@@ -297,7 +297,9 @@ private:
     AutoClimbControl update_manual_support_control(const rmcs_msgs::Keyboard& keyboard) {
         AutoClimbControl control;
 
-        if (keyboard.b) {
+        if (keyboard.b || *rotary_knob_switch_ == rmcs_msgs::Switch::UP) {
+            manual_support_retracting_ = false;
+            manual_support_retract_block_count_ = 0;
             back_climber_zero_velocity_hold_ = false;
             control.back_climber_velocity = climber_back_control_velocity_abs_;
             return control;
@@ -710,8 +712,8 @@ private:
 
     std::shared_ptr<ChassisClimberFrontPowerLimiter> front_power_limiter_;
 
-    double back_climber_retract_first_torque_ = 8.0;
-    double back_climber_retract_second_torque_ = 0.5;
+    double back_climber_retract_first_torque_ = 10.0;
+    double back_climber_retract_second_torque_ = 1.0;
     int back_climber_recover_count = 0;
 };
 } // namespace rmcs_core::controller::chassis
