+++ rmcs_ws/src/rmcs_core/src/controller/chassis/hero_chassis_controller.cpp
@@ -189,6 +189,8 @@ public:
             angular_velocity *=
                 std::clamp(measured_translational_speed / translational_velocity_max, 0.0, 0.3);
 
+            angular_velocity = 0.0;
+
         } break;
         case rmcs_msgs::ChassisMode::SPIN: {
             angular_velocity =
